python_Project_Managment/ProjectManagementFastAPI/backend/requirements.txt

fastapi
uvicorn
sqlalchemy
python-jose
passlib[bcrypt]
python-multipart
